Initialize movie list on application create
